requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.110.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.6
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.45.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pycparser==2.21
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
stripe==8.5.0
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.27.1
websockets==13.0.1
yarl==1.11.1
